pyproject.toml
[project]
name = "vcompany"
version = "0.1.0"
description = "Autonomous Multi-Agent Development System"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.6",
    "pydantic>=2.11,<3",
    "pydantic-settings>=2.13,<3",
    "pyyaml>=6.0",
    "libtmux>=0.55.0,<0.56",
    "jinja2>=3.1",
    "rich>=14.0",
    "discord-py>=2.7,<3",
]

[project.scripts]
vco = "vcompany.cli.main:cli"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "ruff>=0.9",
    "pytest-asyncio>=0.24",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vcompany"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"
